# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoMesh Studio",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "summary": "Projects",
        "operationId": "projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectInfo"
                  },
                  "type": "array",
                  "title": "Response Projects Api Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Project",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}": {
      "get": {
        "summary": "Project",
        "operationId": "project_api_projects__pid__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Project",
        "operationId": "patch_project_api_projects__pid__patch",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/views/{name}": {
      "post": {
        "summary": "Upload View",
        "operationId": "upload_view_api_projects__pid__views__name__post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_view_api_projects__pid__views__name__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "View Settings",
        "operationId": "view_settings_api_projects__pid__views__name__put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete View",
        "operationId": "delete_view_api_projects__pid__views__name__delete",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/views/{name}/image": {
      "get": {
        "summary": "Prepared View",
        "operationId": "prepared_view_api_projects__pid__views__name__image_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/views/{name}/mask": {
      "put": {
        "summary": "Save Mask",
        "operationId": "save_mask_api_projects__pid__views__name__mask_put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_save_mask_api_projects__pid__views__name__mask_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/views/{name}/alignment": {
      "get": {
        "summary": "Alignment Image",
        "operationId": "alignment_image_api_projects__pid__views__name__alignment_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "side",
                "front",
                "rear",
                "top"
              ],
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/jobs": {
      "get": {
        "summary": "Jobs",
        "operationId": "jobs_api_projects__pid__jobs_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobInfo"
                  },
                  "title": "Response Jobs Api Projects  Pid  Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Queue Job",
        "operationId": "queue_job_api_projects__pid__jobs_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jid}": {
      "get": {
        "summary": "Job",
        "operationId": "job_api_jobs__jid__get",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jid}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_jobs__jid__cancel_post",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jid}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_jobs__jid__events_get",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "default": "Untitled vehicle"
          },
          "mode": {
            "type": "string",
            "enum": [
              "blueprint",
              "photo"
            ],
            "title": "Mode",
            "default": "blueprint"
          }
        },
        "type": "object",
        "title": "ProjectCreate"
      },
      "ProjectInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "blueprint",
              "photo"
            ],
            "title": "Mode"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created": {
            "type": "string",
            "title": "Created"
          },
          "updated": {
            "type": "string",
            "title": "Updated"
          },
          "last_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Job Id"
          },
          "thumbnail_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Id"
          },
          "views": {
            "items": {
              "$ref": "#/components/schemas/ViewInfo"
            },
            "type": "array",
            "title": "Views"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "mode",
          "config",
          "revision",
          "created",
          "updated",
          "last_job_id",
          "views"
        ],
        "title": "ProjectInfo"
      },
...
    }
  }
}